clean up *& authorisation

+++ app/Http/Controllers/dashboard/AdminArticlesController.php
@@ -8,6 +8,7 @@ use App\Models\Article;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
 use App\Http\Requests\CreateArticleRequest;
 use App\Http\Requests\UpdateArticleRequest;
@@ -16,7 +17,7 @@ class AdminArticlesController extends Controller
 {
     public function __construct()
     {
-        $this->middleware(['auth'])->only(['store', 'destroy','edit']);
+        $this->middleware(['auth']);
     }
     public function index()
     {
@@ -42,9 +43,11 @@ class AdminArticlesController extends Controller
         // $request->cover_image->move(public_path('images/articles'),$cover_image);
        // }
        $cover_image='images/articles'.'/'.time().'.'.$request->cover_image->extension();
+
        $request->cover_image->move(public_path('images/articles'),$cover_image);
 
         $slug=Str::slug($request->title,'-');
+
         $request->user()->article()->create([
             'title'=>$request->title,
             'slug'=>$slug,
@@ -55,7 +58,7 @@ class AdminArticlesController extends Controller
             'body'=>$request->content,
             'views'=>rand(5,200),
         ]);
-        return redirect(route('admin.articles.index'))->with('success', 'Article  created successfully!');
+        return redirect(route('admin.articles.manage'))->with('success', 'Article  created successfully!');
     }
 
     public function show($id)
@@ -63,22 +66,33 @@ class AdminArticlesController extends Controller
         //
     }
 
-    public function edit(Article $article)
+    public function edit($id)
     {
+
         //
         //dd($article);
         //$article=Article::findOrFail($article);
+        $article=Article::findOrFail($id);
+if($article->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
 
         return view('admin.articles.edit',[
             'article'=>$article,
         ]);
     }
 
-    public function update(UpdateArticleRequest $request,Article $article)
+    public function update(UpdateArticleRequest $request,$id)
     {
+         $article=Article::findOrFail($id);
+        if($article->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $request->validated();
         //
 
+       // dd($article);
         $current_cover_image=$article->image();
 
         if ($request->hasFile('cover_image')) {
@@ -93,9 +107,11 @@ class AdminArticlesController extends Controller
             $cover_image=$article->image();
         }
 
-
         $slug=Str::slug($request->title,'-');
-        $request->user()->article()->update([
+
+
+
+        $article->update([
             'title'=>$request->title,
             'slug'=>time().'-'.$slug,
             'min_to_read'=>rand(5,15),
@@ -105,20 +121,28 @@ class AdminArticlesController extends Controller
             'body'=>$request->content,
             'views'=>rand(5,200),
         ]);
-        return redirect(route('admin.articles.index'))->with('success', 'Article  updated successfully!');
+        return redirect(route('admin.articles.manage'))->with('success', 'Article  updated successfully!');
 
     }
 
-    public function destroy(Article $article)
+    public function destroy($id)
     {
+         $article=Article::findOrFail($id);
+        if($article->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         //
+
         $current_cover_image=$article->image();
         if(File::exists($current_cover_image))
         {
             File::delete($current_cover_image);
         }
 
+
         $article->delete();
-        return redirect(route('admin.articles.index'))->with('danger', 'Article  Deleted!');
+        return redirect(route('admin.articles.manage'))->with('danger', 'Article  Deleted!');
     }
+
+
 }
